moved models from params to a separate file
