+++ pushsdkandroid/src/main/java/com/push/android/pushsdkandroid/core/PushKApi.kt
@@ -4,6 +4,10 @@ import android.content.Context
 import com.push.android.pushsdkandroid.add.Answer
 import com.push.android.pushsdkandroid.add.GetInfo
 import com.push.android.pushsdkandroid.logger.PushKLoggerSdk
+import com.push.android.pushsdkandroid.models.PushKDataApi
+import com.push.android.pushsdkandroid.models.PushKDataApi2
+import com.push.android.pushsdkandroid.models.PushKFunAnswerGeneral
+import com.push.android.pushsdkandroid.models.PushKFunAnswerRegister
 import java.io.BufferedReader
 import java.io.DataOutputStream
 import java.io.InputStreamReader
